api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.1
requests==2.31.0
python-jose==3.3.0
python-multipart==0.0.6
